Revert "Merge branch 'cm-14.1' of https://github.com/CyanogenMod/android_frameworks_base into OMS7-N"
This reverts commit 64b383f8f8769ec0c9a661f3da68f71ecc12ca0e, reversing
changes made to a6ed530b2f612b7f51d39de24c2220b94a259209.

+++ packages/ExternalStorageProvider/src/com/android/externalstorage/ExternalStorageProvider.java
@@ -466,7 +466,10 @@ public class ExternalStorageProvider extends DocumentsProvider {
         displayName = FileUtils.buildValidFatFilename(displayName);
 
         final File before = getFileForDocId(docId);
-        final File after = FileUtils.buildUniqueFile(before.getParentFile(), displayName);
+        final File after = new File(before.getParentFile(), displayName);
+        if (after.exists()) {
+            throw new IllegalStateException("Already exists " + after);
+        }
         if (!before.renameTo(after)) {
             throw new IllegalStateException("Failed to rename to " + after);
         }
@@ -563,7 +566,6 @@ public class ExternalStorageProvider extends DocumentsProvider {
             throws FileNotFoundException {
         final MatrixCursor result = new MatrixCursor(resolveDocumentProjection(projection));
 
-        query = query.toLowerCase();
         final File parent;
         synchronized (mRootsLock) {
             parent = mRoots.get(rootId).path;
